Fix repository tests
