+++ app/src/test/java/com/marekmacko/kotlinapp/repository/RemoteWeatherRepositoryTest.kt
@@ -5,12 +5,10 @@ import com.marekmacko.kotlinapp.RxImmediateSchedulerRule
 import com.marekmacko.kotlinapp.api.WeatherService
 import com.marekmacko.kotlinapp.data.response.WeeklyForecast
 import com.marekmacko.kotlinapp.data.ui.ForecastShort
-import com.nhaarman.mockito_kotlin.any
-import com.nhaarman.mockito_kotlin.whenever
+import com.nhaarman.mockito_kotlin.*
 import com.tngtech.java.junit.dataprovider.DataProviderRunner
 import com.tngtech.java.junit.dataprovider.UseDataProvider
 import io.reactivex.Observable
-import io.reactivex.observers.TestObserver
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -30,40 +28,36 @@ class RemoteWeatherRepositoryTest {
     @JvmField
     val mockitoRule: MockitoRule = MockitoJUnit.rule()
 
-    @Mock lateinit var weatherService: WeatherService
+    @Mock
+    lateinit var weatherService: WeatherService
 
-    @InjectMocks lateinit var remoteWeatherRepository: RemoteWeatherRepository
+    @InjectMocks
+    lateinit var weatherRepository: WeatherRepository
+
+    private val onLoadMock = mock<(List<ForecastShort>) -> Unit>()
+    private val onErrorMock = mock<(String) -> Unit>()
 
     @Test
     @UseDataProvider("getWeeklyForecast", location = arrayOf(DataProviderSource::class))
     fun getWeeklyForecastReturnData(weeklyForecast: WeeklyForecast) {
         whenever(weatherService.getWeeklyForecast(any())).thenReturn(Observable.just(weeklyForecast))
+        whenever(onLoadMock.invoke(any())).thenReturn(any())
 
-        val observableResult = remoteWeatherRepository.getWeeklyForecast()
-        val testObserver = TestObserver<List<ForecastShort>>()
-        observableResult.subscribe(testObserver)
-        testObserver.assertComplete()
-        testObserver.assertNoErrors()
-        testObserver.assertValueCount(1)
-        testObserver.assertValue { // TODO
-            it[0].description == DataProviderSource.DESCRIPTION
-        }
-        testObserver.assertValue {
-            it[0].minTemperature == DataProviderSource.MIN_TEMPERATURE
-        }
-        testObserver.assertValue {
-            it[0].maxTemperature == DataProviderSource.MAX_TEMPERATURE
-        }
+        weatherRepository.getWeeklyForecast(onLoadMock, onErrorMock)
+
+        verify(onLoadMock, times(1)).invoke(any())
     }
 
     @Test
-    fun getWeeklyForecastReturnError() {
-        val error = Throwable()
+    @UseDataProvider("getWeeklyForecastShort", location = arrayOf(DataProviderSource::class))
+    fun getWeeklyForecastReturnError(forecasts: List<ForecastShort>) {
+        val errorMessage = "error"
+        val error = Throwable(errorMessage)
         whenever(weatherService.getWeeklyForecast(any())).thenReturn(Observable.error(error))
+        whenever(onErrorMock.invoke(any())).thenReturn(any())
 
-        val observableResult = remoteWeatherRepository.getWeeklyForecast()
-        val testObserver = TestObserver<List<ForecastShort>>()
-        observableResult.subscribe(testObserver)
-        testObserver.assertError(error)
+        weatherRepository.getWeeklyForecast(onLoadMock, onErrorMock)
+
+        verify(onErrorMock, times(1)).invoke(any())
     }
 }
